# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Company Managment",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/token": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get the Bearer Token",
                "description": "Return bearer token",
                "operationId": "getBearerToken",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "Content Type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "description": "Request body is required",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "email": "admin@admin.com",
                                    "password": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully generated token",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/companies/": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get list of Companies",
                "description": "Returns a list of Companies",
                "operationId": "getListOfCompanies",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "Content Type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Company objects",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employees/": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get list of Employees",
                "description": "Returns a list of Employees",
                "operationId": "getListOfEmployees",
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "Content Type",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Company objects",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}